# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Deriv Quant Research & Execution Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /trade/history:
    get:
      operationId: getTradeHistory
      tags: [trades]
      summary: Get historical trades
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: symbol
          in: query
          schema:
            type: string
        - name: strategy
          in: query
          schema:
            type: string
        - name: mode
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Trade history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Trade"
...
components:
  schemas:
...
    Trade:
      type: object
      properties:
        id:
          type: integer
        brokerTradeId:
          type: ["string", "null"]
        symbol:
          type: string
        strategyName:
          type: string
        side:
          type: string
          enum: [buy, sell]
        entryTs:
          type: string
        exitTs:
          type: ["string", "null"]
        entryPrice:
          type: number
        exitPrice:
          type: ["number", "null"]
        sl:
          type: number
        tp:
          type: number
        size:
          type: number
        pnl:
          type: ["number", "null"]
        status:
          type: string
          enum: [open, closed, cancelled]
        mode:
          type: string
          enum: [paper, live]
        notes:
          type: ["string", "null"]
        confidence:
          type: ["number", "null"]
        trailingStopPct:
          type: ["number", "null"]
        peakPrice:
          type: ["number", "null"]
        maxExitTs:
          type: ["string", "null"]
        exitReason:
          type: ["string", "null"]
        currentPrice:
          type: ["number", "null"]
      required:
        - id
        - brokerTradeId
        - symbol
        - strategyName
        - side
        - entryTs
        - exitTs
        - entryPrice
        - exitPrice
        - sl
        - tp
        - size
        - pnl
        - status
        - mode
        - notes
        - confidence
        - trailingStopPct
        - peakPrice
        - maxExitTs
        - exitReason
        - currentPrice
...